ADDED: Allow application initiated session management.
There is a new session option called create that can be auto (default)
or noauto.  In the latter case sessions are not created automatically,
but must be created explicitely by the user using http_open_session/2.

Based on a patch by Matt Lilley with the same objective.  This version
tries to simply writing code that works in both session models.