Support DELETE verb in http_open/3
As requested by Wouter Beek, this patch adds support for HTTP 1.1 verb
DELETE.  Its implementation is very similar to GET and HEAD.

I actually needed this for a project I'm working.  Without this patch,
I got this:

    ?- http_open('http://127.0.0.1:3000/foo', Stream, [method(delete)]),
    |    close(Stream).
    ERROR: Domain error: `method' expected, found `delete'

With this patch, I get:

    Stream = <stream>(0x7fb71212e070).

and the correct request is made to the server.