Revert "MODIFIED: http_open/3: process Content-encoding as Transfer-encoding."
This reverts commit 9c6b6bdba91aa9bc978842f6bc9b89e914ea42f8.

transfer_encoding/2 should only yield the transfer encoding. The problem
that arose with this change is if the content-encoding is correctly
used to indicate for example a gzip file. We must not decompress that
transparently. Instead, treat content-encoding as transfer-encoding
only if the content-type does not indicate a compressed file.
Patch follows.