ENHANCED: HTTP Unix daemon: Read certificate and key before privileges are dropped.
Note that these changes restrict the format of the private key to what
load_private_key/3 can handle (PEM-encoding), which is the most
commonly encountered encoding. Overall, these changes simplify
maintenance of the daemon as well as of user scripts that need to
access files which are only readable with root privileges. If more key
formats are needed, load_private_key/3 should be extended.