MODIFIED: HTTP Unix daemon no longer uses https/server.crt by default.
This default was undocumented and, more importantly, prevented running
a server that relies entirely on SNI to present certificates.

No longer requiring a mandatory certificate/key pair to be specified
allows you to automatically distinguish clients that support SNI: If
you omit --certfile, clients that do not support SNI are rejected.