FIXED: http_open/3 with keep-alive connections
Fixes two issues: (1) streams may be left open (leaking stream handles)
when an error occurs during the HTTP handhake and (2) errors during the
http handshake were propagated without the error(_,_) wrapper.    With
http handshake we refer to sending and reading the header as well as, for https,
the TLS handshake.